requirements.txt
africastalking==1.2.3
certifi==2020.6.20
chardet==3.0.4
contextlib2==0.6.0.post1
idna==2.10
python-dotenv==0.14.0
requests==2.24.0
schema==0.7.3
urllib3==1.25.10
